refactor: 돈 지불하고 구매 로또 개수 정하는 메서드 분리
